requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-chess==0.24.1
PyYAML>=4.2b1
requests>=2.21.0
urllib3==1.24.2
backoff==1.7.1
pytest
pytest-cov